refactor: mv util_enum -> util
